ddp changes

+++ trainer/early_stopping.py
@@ -4,7 +4,8 @@ import os
 
 class EarlyStopping:
     """Early stops the training if validation score doesn't improve after a given patience."""
-    def __init__(self, model, model_type,optimizer, config, checkpoint_dir, mnt_mode = 'max', patience=7, verbose=False, delta=0, trace_func=print,parallel_type=None):
+    def __init__(self, model, model_type,optimizer, config, checkpoint_dir, mnt_mode = 'max', patience=7, 
+                 verbose=False, delta=0, trace_func=print,parallel_type=None, device = 0):
         self.patience = patience
         self.verbose = verbose
         self.counter = 0
@@ -18,6 +19,7 @@ class EarlyStopping:
         self.config = config
         self.optimizer = optimizer
         self.parallel_type = parallel_type
+        self.device = device
     
         self.trace_func = trace_func
     def __call__(self, score, epoch, model):
@@ -38,28 +40,29 @@ class EarlyStopping:
 
     def save_checkpoint(self, score,epoch, model):
         '''Saves model when validation metric decrease.'''
-        if self.verbose:
-            self.trace_func(f'Metric improved ({self.best_score:.6f} --> {score:.6f}).  Saving model ...')
-        if self.parallel_type is not None:
-            state = {
-            'arch': self.model_type,
-            'epoch': epoch,
-            'state_dict': self.model.module.state_dict(),
-            'optimizer': self.optimizer.state_dict(),
-            'best_score': self.best_score,
-            'config': self.config
-        }
-        else:
-            state = {
+        if self.device == 0:
+            if self.verbose:
+                self.trace_func(f'Metric improved ({self.best_score:.6f} --> {score:.6f}).  Saving model ...')
+            if self.parallel_type is not None:
+                state = {
                 'arch': self.model_type,
                 'epoch': epoch,
-                'state_dict': self.model.state_dict(),
+                'state_dict': self.model.module.state_dict(),
                 'optimizer': self.optimizer.state_dict(),
                 'best_score': self.best_score,
                 'config': self.config
             }
-        previous_checkpoint_pth = os.path.join(self.checkpoint_dir,f"checkpoint-epoch{epoch-1}.pth")
-        if os.path.exists(previous_checkpoint_pth):
-            os.remove(previous_checkpoint_pth)
-        checkpoint_path = os.path.join(self.checkpoint_dir,f"checkpoint-epoch{epoch}.pth")
-        torch.save(state, checkpoint_path)
+            else:
+                state = {
+                    'arch': self.model_type,
+                    'epoch': epoch,
+                    'state_dict': self.model.state_dict(),
+                    'optimizer': self.optimizer.state_dict(),
+                    'best_score': self.best_score,
+                    'config': self.config
+                }
+            previous_checkpoint_pth = os.path.join(self.checkpoint_dir,f"checkpoint-epoch{epoch-1}.pth")
+            if os.path.exists(previous_checkpoint_pth):
+                os.remove(previous_checkpoint_pth)
+            checkpoint_path = os.path.join(self.checkpoint_dir,f"checkpoint-epoch{epoch}.pth")
+            torch.save(state, checkpoint_path)
